requirements.txt
# Gravitational Wave Hunter v2.0 Requirements

# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
torch>=1.9.0
scikit-learn
seaborn>=1.0.0
h5py>=3.0.0

# Data processing
pywt>=1.1.0
matplotlib>=3.4.0
pandas>=1.3.0

# Configuration and utilities
PyYAML>=6.0
requests>=2.25.0
pathlib2>=2.3.0

# Development and testing
pytest>=6.0.0
pytest-cov>=2.12.0
black>=21.0.0
flake8>=3.9.0

# Optional: For real GWOSC data access
gwosc>=0.8.0  # For accessing GWOSC event information
gwpy>=3.0.0  # For fetching open data with science-mode segments
# ligo-gracedb>=2.0.0  # Uncomment for GraceDB access
